Functionally ready
Created a layout for all pages with navigation, info and links to github, linkedin and shadertoy.
Created second files for the website in english. Not the best idea for scaling but I don't think I'll add much more.
The blog is now working, just need to add the file to the blog and blog_en folders.